fix qavan entry in German + other
The user comment was right about "ich grüße dich/dir". But I decided to fix it differently, because "grüße" is "greeting", but not "salute". So I changed that to "salutieren" which is then correct using "dir" (it's possible that's the reson it was wrong in the first place)