xml2json.py: split search tags into array of tags in JSON dump
This makes it easier to determine how closely a query matches an
individual tag, as opposed to the whole search tags string. This
sort of change would normally necessitate bumping the format version
field, but this field was inadvertently not being exported (this
change fixes that as well), and in any case, the JSON format of the
database isn't actually used anywhere, yet.

A similar change was considered for the "synonyms", "antonyms",
"see_also", and "components" fields, but these fields did not
follow a consistent enough format to make treating them as arrays
of references to other entries practical.