Decompose Unicode characters
Normalize Unicode to the NFKD form (NFD is probably fine, but NFKD is
chosen out of an abundance of caution) in the JSON representation of
the database so that the app doesn't have to do it at load time.

As part of this, change the shebang to explicitly request python 3,
since python 2 Unicode support is hard. Prior to this change, this
script worked in both Python 2 and Python 3.

This is work towards dadap/flingon-assister#19