Adjust possessive suffix translations
The possessive suffix -raj, which is glossed as "your (plural)" in
English, was translated as if the possessed item were plural, rather
than the possessor. Fix this translation to make the possessor plural,
and also adjust all of the possessive suffix translations to show
that the possessed may be singular or plural.