jonta' -> en: engine -> de: Antrieb, Motor
jonta' is currently translated as "Maschine" in german - while not wrong, it's misleading. "Maschine" literally means "machine", which could be anything from a coffe-maker to a robot.
I think a better translation might be "Antrieb, Motor", although those may be misleading as well ("propulsion system", "motor").
Maybe add as note that in the movie it referred to the spaceship's propulsion?