Note on avoiding formal "Sie"
I've added a note for German editors to avoid the formal "Sie" ("you") because it looks pretty weird in such an app. Not because it's a Klingon app, but just in general and in digital Apps, formal "Sie" is usually not used. You would encounter it on government pages or in online sores maybe. To avoid any discussion if Du or Sie is better, I suggest using a general statement avoiding the direct address of the reader.